+++ Pop-Grupa-A-Frontend/src/commons/ApiLinks.js
@@ -1,0 +1,8 @@
+const baseUrl = "http://localhost:5000/"
+
+const taskUrl = `${baseUrl}task/` // plus app
+
+
+export{
+    taskUrl
+}
